add to windows build
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@6545 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/event_handlers/mod_cdr_csv/mod_cdr_csv.c
@@ -129,8 +129,8 @@ static void write_cdr(const char *path, const char *log_line)
 	}
 
 	if (fd) {
-		size_t bytes_in, bytes_out;
-		bytes_out = strlen(log_line);
+		unsigned int bytes_in, bytes_out;
+		bytes_out = (unsigned)strlen(log_line);
 		
 		if (fd->bytes + bytes_out > UINT_MAX) {
 			do_rotate(fd);
